==> scripts/9-var-genes-sc3-k29.Rmd <==
---
title: "Gene sets for SC3 clusters"
author: "Amy Gill"
date: "4/21/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(Seurat)
library(tidyverse)

load("tnbc_seurat_sc3.rda")
```


```{r}
tnbc_seurat <- SetIdent(tnbc_seurat, value = "sc3_29_clusters")
sc3_k29_markers <- FindAllMarkers(tnbc_seurat, logfc.threshold = 0.5, slot = "scale.data")

# keep only differentially expressed markers with an adjusted pval less than .05
sc3_k29_markers <- sc3_k29_markers %>%
	filter(p_val_adj <= .05)

save(sc3_k29_markers, file = "sc3_k29_markers.rda")
```

```{r}
sc3_k29_markers
```


```{r}
# number of differentially expressed genes for each cluster
sc3_k29_markers %>%
	group_by(cluster) %>%
	summarize(n = n())
```

```{r}
top3 <- sc3_k29_markers %>%
	group_by(cluster) %>%
	top_n(3, wt = avg_diff)

DoHeatmap(tnbc_seurat, features = top3$gene) + NoLegend()
```




```{r}
top10 <- sc3_k29_markers %>%
	group_by(cluster) %>%
	top_n(10, wt = avg_diff)

DoHeatmap(tnbc_seurat, features = top10$gene) + NoLegend() +
	theme(axis.text.y = element_blank())
```


Some clusters seem composed of cells with unique expression patterns of the top 10 differentially expressed genes identified for each cluster. (for example, 6, 26, 28, 29). Other groups of distinct clusters seem to have highly similar patterns of differential expression (for example, 9-10 and 8-19-20), suggesting this approach may not be returning the most parsimonious group of biologically relevant clusters. Still other clusters had few or no distinguishing expression patterns across the set of top 10 differentially expressed genes per cluster (for example, 1, 3, 17, 18). Additional clusters had an extremely small number of cells (cluster 2 = 1 cell, cluster 5 = 5 cells, cluster 16 = 6 cells, cluster 22 = 8 cells) and may not reflect biologically meaningful subsets or may represent populations that are too rare to characterize with statistical confidence. Combined, these results suggest that the estimated optimal value of $k$ from SC3 may result in overly fine resolution of clusters and that the true number of clusters is less than 29.
