==> scripts/14-mia-cells-seurat-sc3.Rmd <==
---
title: "Seurat and cell type MIA for cell cluster membership"
author: "Amy Gill"
date: "4/21/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(Seurat)
library(RColorBrewer)
library(pheatmap)

load("tnbc_seurat_sc3.rda")

# set color palettes for heatmap
reds <- colorRampPalette(c("white", "red"))(100)
cools_rev <- colorRampPalette(c("yellow", "limegreen", "black"))(100)
```

```{r}
# create vectors of celltype and cluster to iterate over
sc3_names <- names(table(tnbc_seurat$sc3_29_clusters))
seurat_names <- names(table(tnbc_seurat$seurat_clusters)) 

# set total number of cells
total_count <- length(tnbc_seurat$seurat_clusters)
```

```{r}
# initialize output matrix
hyper <- matrix(NA, nrow = length(sc3_names), ncol = length(seurat_names))
rownames(hyper) <- sc3_names
colnames(hyper) <- seurat_names
```


```{r}
# calculate all hypergeometric p-values for overlap between cell type and Seurat cluster
for (i in 1:length(sc3_names)) {
	sc3_count <- sum(tnbc_seurat$sc3_29_clusters == sc3_names[i])    # number of cells for given cell type
	for (j in 1:length(seurat_names)) {
		seurat_count <- sum(tnbc_seurat$seurat_clusters == seurat_names[j])    # number of cells in given cluster
		overlap_count <- sum(tnbc_seurat$sc3_29_clusters == sc3_names[i] & tnbc_seurat$seurat_clusters == seurat_names[j])    # number of cells of type in cluster
		
		# calculate p-value for enrichment of cells in groups with hypergeometric test
		hyper_pval <- phyper(overlap_count - 1, sc3_count, total_count - sc3_count, seurat_count, lower.tail = FALSE)
		
		hyper[i, j] <- hyper_pval
	}
}
```

```{r}
hyper
```

```{r}
pheatmap(hyper,
		 cluster_cols = FALSE,
		 cluster_rows = FALSE,
		 color = cools_rev,
		 main = "p-values for overlap of cells in SC3 and Seurat clusters")
```

```{r}
pheatmap(-log10(hyper),
		 cluster_cols = FALSE,
		 cluster_rows = FALSE,
		 color = reds,
		 main = "-log10(p-values) for overlap of cells in SC3 and Seurat clusters")
```

```{r}
pheatmap(sqrt(-log10(hyper)),
		 cluster_cols = FALSE,
		 cluster_rows = FALSE,
		 color = reds,
		 main = "sqrt(-log10(p-values)) for overlap of cells in SC3 and Seurat clusters")
```
Some SC3 clusters correspond well to a single Seurat (Louvain) cluster (for example, SC3-29 and Louvain-7, or SC3-28 and Louvain-8). Some SC3 clusters divide a single Louvain cluster (for example, Louvain-3 is divided into SC3-9/10, and Louvain-0 is divided into various clusters including SC3-11/12). Still other SC3 clusters aggregate cells from a variety of Louvain clusters (for example, SC3-3 is composed of cells from Louvain-2/4/6).
